Implement the /stock command
It works with some tests made but still need some work on i1-chat to configure
the exchanges and queues properly.

The plan for now is to use a direct exchange with the consumers (chat clients)
creating the queues and binding this queue to the routing key with the name of
the room (room1, room2, etc. for simplicity).

Other patterns are being considered and using other patterns such as the topic
or even changing a direct exchange for a header exchange.