Create the exchange and queues on the backend
Also log what is happening on stdout. The idea is to have the backend creating
all the infrastructure on the broker so the clients and the bots can operate
upon.

Things like attributes of the exchanges and queues are made explicit although
somewhat more verbose. This makes it easier to match what's on the bot also.